feat(reservation)

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,5 +1,6 @@
 package nextstep.reservation;
 
+import nextstep.auth.AuthenticationException;
 import nextstep.member.Member;
 import nextstep.member.MemberDao;
 import nextstep.schedule.Schedule;
@@ -25,14 +26,12 @@ public class ReservationService {
         this.memberDao = memberDao;
     }
 
-    public Long create(ReservationRequest reservationRequest) {
+    public Long create(Member member, ReservationRequest reservationRequest) {
+        if (member == null) {
+            throw new AuthenticationException();
+        }
         Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
         if (schedule == null) {
-            throw new NullPointerException();
-        }
-
-        Member member = memberDao.findById(reservationRequest.getMemberId());
-        if (member == null) {
             throw new NullPointerException();
         }
 
@@ -58,10 +57,14 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Member member, Long id) {
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
             throw new NullPointerException();
+        }
+
+        if (!reservation.sameMember(member)) {
+            throw new AuthenticationException();
         }
 
         reservationDao.deleteById(id);
